feat: moving away from having the utils on utils package and place them on BTC folder
