+++ packages/bitcoin/package.json
@@ -61,6 +61,8 @@
   },
   "dependencies": {
     "@buildwithsygma/core": "workspace:^",
+    "bip32": "4.0.0",
+    "bip39": "3.1.0",
     "bitcoinjs-lib": "^6.1.6"
   }
 }
