--- packages/utils/src/bitcoin/index.ts
@@ -1,2 +1,0 @@
-export * from './blockstream.js';
-export * from './wallet.js';
