+++ packages/utils/src/index.ts
@@ -1,2 +1,1 @@
 export * from './liquidity.js';
-export * from './bitcoin/index.js';
